adapt cpp tests
